pyproject.toml
[project]
name = "Simple PDF RAG"
version = "1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.20",
    "faiss-cpu>=1.9.0.post1",
    "haystack-ai>=2.6.1",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.3",
    "langchain-experimental>=0.3.3",
    "langchain-groq>=0.2.0",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.2.3",
    "langchain>=0.3.4",
    "langgraph>=0.2.39",
    "libmagic>=1.0",
    "matplotlib>=3.9.2",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.3.0",
]
